moved json files to own folder, added footer component
