+++ src/app/resourceLinks/allLinks.ts
@@ -1,18 +1,18 @@
-import mdLinks from './markdownLinks.json';
-import cliLinks from './cliLinks.json';
-import htmlLinks from './htmlLinks.json';
-import cssLinks from './cssLinks.json';
-import jsLinks from './jsLinks.json';
-import jQueryLinks from './jQueryLinks.json'
-import webAPILinks from './webAPILinks.json';
-import restAPILinks from './restfulAPILinks.json';
-import cssFrameworkLinks from './cssFrameworkLinks.json';
-import nodeJSLinks from './nodeJSLinks.json';
-import oopLinks from './oopLinks.json';
-import expressLinks from './expressJSLinks.json';
-import mysqlLinks from './mysqlLinks.json';
-import mongodbLinks from './mongodbLinks.json';
-import etcLinks from './etcLinks.json';
+import mdLinks from './json/markdownLinks.json';
+import cliLinks from './json/cliLinks.json';
+import htmlLinks from './json/htmlLinks.json';
+import cssLinks from './json/cssLinks.json';
+import jsLinks from './json/jsLinks.json';
+import jQueryLinks from './json/jQueryLinks.json'
+import webAPILinks from './json/webAPILinks.json';
+import restAPILinks from './json/restfulAPILinks.json';
+import cssFrameworkLinks from './json/cssFrameworkLinks.json';
+import nodeJSLinks from './json/nodeJSLinks.json';
+import oopLinks from './json/oopLinks.json';
+import expressLinks from './json/expressJSLinks.json';
+import mysqlLinks from './json/mysqlLinks.json';
+import mongodbLinks from './json/mongodbLinks.json';
+import etcLinks from './json/etcLinks.json';
 
 
 const allLinks = [
